% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{bind_rows_with_na}
\alias{bind_rows_with_na}
\title{Row bind columns with different headers}
\usage{
bind_rows_with_na(df_list)
}
\arguments{
\item{df_list}{list of data.frames to rowbind}
}
\description{
Row bind columns with different headers
}
