% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\name{check_msl_fields}
\alias{check_msl_fields}
\title{Check if MSL has been merged with holdings file}
\usage{
check_msl_fields(x)
}
\arguments{
\item{x}{data.frame with holdings data}
}
\value{
TRUE if MSL has been merged, FALSE if not
}
\description{
Check if MSL has been merged with holdings file
}
