% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{clean_asset_bench_rf}
\alias{clean_asset_bench_rf}
\title{Handle combing and cleaning asset, rf, and benchmarks}
\usage{
clean_asset_bench_rf(
  x,
  b = NULL,
  rf = NULL,
  freq = NULL,
  date_start = NULL,
  date_end = NULL,
  eps = 0.05,
  add_bench_nm = TRUE,
  add_rf_nm = TRUE
)
}
\arguments{
\item{x}{xts object of asset(s)}

\item{b}{xts object of benchmark(s)}

\item{rf}{optional xts object of risk-free time-series}

\item{freq}{string return frequency: days, weeks, months, quarters, years,
see details}

\item{date_start}{date representing first date. See details.}

\item{date_end}{date representing most recent date. See details.}

\item{eps}{percent of missing values (NAs) to allow, default is 0.05 for 5%.
See details.}
}
\value{
list with $x for asset(s), $b for benchmark(s), $xb for assets and
 benchmark, and $rf if risk-free is entered.
}
\description{
Handle combing and cleaning asset, rf, and benchmarks
}
\details{
Utility function to handle common task of needing to line up asset, bench,
  and rf returns. If left `NULL` the `date_start` and `date_end` will
  truncate the data at the most recent common start date and oldest common
  end date. If further truncation is desired dates can be entered to cut the
  time-series accordingly. `eps` sets the bar for how many missing values
  are allowed per column. The default is 5%, if a column has more than 5%
  of the observations are missing the column is removed. Otherwise the
  missing values are filled with zero. If frequency is entered returns will
  be changed to desired frequency, otherwise leave `NULL` to not alter.
}
\examples{
data(assets)
b <- x[, 1]
x <- x[, -1]
res <- clean_asset_bench_rf(x, b, rf = NULL, freq = "months",
  date_start = "2020-01-01")
head(res$x)
}
