% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{clean_by_col}
\alias{clean_by_col}
\title{Clean Returns by Column}
\usage{
clean_by_col(x, eps = 0.05)
}
\arguments{
\item{x}{return xts}

\item{eps}{% of missing returns to tolerate, default is 0.05 for 5%}
}
\value{
list with $ret for cleaned returns and $miss for any columns
  that had to be removed (if any)
}
\description{
Handle missing returns in each column
  without truncating to common periods for the return matrix. If the
  % of missing values (once data starts in each column) is over the eps
  the column will be removed, otherwise missing values will be replaced
  with zeros.
}
