% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{clean_ids}
\alias{clean_ids}
\title{Clean CUSIPs in holdings tables}
\usage{
clean_ids(tbl_hold)
}
\arguments{
\item{tbl_hold}{data.frame with holdings}
}
\value{
tbl_hold with cleaned up CUSIPs
}
\description{
CUSIPs with all zeros need to be read as NA
}
