% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{co_fundamental_data}
\alias{co_fundamental_data}
\title{Download equity financial data}
\usage{
co_fundamental_data(
  api_keys,
  bucket,
  ids = NULL,
  yrs_back = 1,
  dtype = c("PE", "PB", "PFCF", "DY", "ROE", "MCAP")
)
}
\arguments{
\item{api_keys}{list with api keys}

\item{bucket}{s3 filesystem}

\item{ids}{leave `NULL` to get for all stocks, or enter specific ids}

\item{yrs_back}{how many years of data to pull}

\item{dtype}{data type: one of PE, PB, PFCF, DY, ROE, and MCAP}
}
\value{
does not return data, updates database with fundamental data
}
\description{
Download equity financial data
}
\seealso{
\code{\link{download_fs_formula}}
}
