% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{create_ids}
\alias{create_ids}
\title{Get IDs from multiple fields in Master Security List}
\usage{
create_ids(tbl_msl)
}
\arguments{
\item{tbl_msl}{data.frame containing MSL}
}
\value{
vector of ids
}
\description{
will search (in order): Isin, Sedol, Lei, Cusip, Ticker,
  and Identifier. For use with Factset downloads requiring these types of
  IDs.
}
