% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{cut_time}
\alias{cut_time}
\title{Truncate time-series with start and end dates}
\usage{
cut_time(x, date_start = NULL, date_end = NULL)
}
\arguments{
\item{x}{xts object}

\item{date_start}{optional beginning date to truncate time-series}

\item{date_end}{optional ending date to truncate time-series}
}
\value{
truncated xts
}
\description{
Truncate time-series with start and end dates
}
\details{
If neither date_start or date_end are entered the original
time series, x, will be returned. The date_start and date_end
parameters are inclusive, i.e., the date_start will be the first
date in the time-series. Note dates can be Date objects or "YYYY-MM-DD"
strings.
}
\examples{
data(assets)
# cut time-series to start in Feb 1, 2014 and end on June 30, 2014
x_cut <- cut_time(x, '2014-02-01', '2014-06-30')
head(x)
tail(x)
# leave beginning of time-serie as is, end on June 30, 2014
cut_time(x, NULL, '2014-06-30')
tail(x)
}
