% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{dataframe_to_xts}
\alias{dataframe_to_xts}
\title{Convert dataframe to xts}
\usage{
dataframe_to_xts(xdf)
}
\arguments{
\item{xdf}{data.frame with dates in first column}
}
\value{
xts
}
\description{
Convert dataframe to xts
}
\examples{
data(assets)
my_df <- xts_to_dataframe(x)
my_xts <- dataframe_to_xts(my_df)
head(my_df)
head(my_xts)
}
