% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_fred}
\alias{download_fred}
\title{Download Econ Time-series from FED Database}
\usage{
download_fred(series_id, fred_api)
}
\arguments{
\item{series_id}{character to represent the series you want to download}

\item{fred_api}{API key, see https://fred.stlouisfed.org/ to create one}
}
\value{
xts of economic time-series
}
\description{
Download Econ Time-series from FED Database
}
