% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_tiingo_csv}
\alias{download_tiingo_csv}
\title{Download price time-series from Tiingo}
\usage{
download_tiingo_csv(ticker, t_api, date_start = "1970-01-01", date_end = NULL)
}
\arguments{
\item{ticker}{character of stock ticker to download, see
`download_tiingo_tickers` for downloading multiple tickers}

\item{t_api}{API key}

\item{date_start}{inception for time-series}

\item{date_end}{last day of time-series, if left `NULL` will default to today}
}
\value{
`data.frame` of historical prices (adjusted for splits and dividends)
  with dates in first column
}
\description{
Download price time-series from Tiingo
}
