% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\name{drill_down}
\alias{drill_down}
\title{Drilldown to a lower layer}
\usage{
drill_down(tbl_hold, layer = 1, latest = TRUE)
}
\arguments{
\item{tbl_hold}{data.frame with holdings}

\item{layer}{numeric value for which layer to drilldown to}

\item{latest}{boolean to subset for most recent holdings}
}
\value{
list with underlying holdings data.frame and a data.frame
  for holdings not found in the MSL
}
\description{
Drilldown to a lower layer
}
