% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{extract_list}
\alias{extract_list}
\title{Utility function to extract a named field from a list
 while returning NA instead of NULL values for missing data}
\usage{
extract_list(x, nm)
}
\arguments{
\item{x}{list}

\item{nm}{vector of names (characters)}
}
\value{
values in list that correspond to the named field
}
\description{
Utility function to extract a named field from a list
 while returning NA instead of NULL values for missing data
}
