% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{fill_ix}
\alias{fill_ix}
\title{Utility function to fill missing values}
\usage{
fill_ix(a, b)
}
\arguments{
\item{a}{starting vector with potentially missing values}

\item{b}{corresponding vector of the same length used to fill parts of a
that are missing values}
}
\value{
a with missing values filled by b
}
\description{
Fills missing values of vector a with vector b, where a
  and b are equal lengths with corresponding values, e.g., Tickers and
  CUSIPS for the same securities
}
\note{
if a and b are not the same length the function will return a as
originally inputted
}
