% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{first_comm_start}
\alias{first_comm_start}
\title{Find intersect start date}
\usage{
first_comm_start(x)
}
\arguments{
\item{x}{xts}
}
\value{
Date representing the common start date of multiple assets.
}
\description{
Find intersect start date
}
\examples{
data(assets)
x[1:3, 1] <- NA
head(x)
first_comm_start(x)
}
