% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\name{flatten_drill_down}
\alias{flatten_drill_down}
\title{Flatten drill down list}
\usage{
flatten_drill_down(res)
}
\arguments{
\item{res}{output from qual_roll_up() function}
}
\value{
adds flat slot with data.frames
}
\description{
Flatten drill down list
}
