% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{get_ids}
\alias{get_ids}
\title{Find IDs in holdings table from multiple fields}
\usage{
get_ids(tbl_hold)
}
\arguments{
\item{tbl_hold}{data.frame with holdings}
}
\value{
vector of IDs
}
\description{
will look for DtcName, Ticker, Cusip, Sedol, Lei, and
  Identifier fields to combine into an "IDs" vector
}
