% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{guess_freq}
\alias{guess_freq}
\title{Guess Frequency with xts periodicity}
\usage{
guess_freq(x)
}
\arguments{
\item{x}{xts}
}
\value{
frequency as string, e.g., "days", "months"
}
\description{
Guess Frequency with xts periodicity
}
\examples{
data(assets)
guess_freq(x)
}
