% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{hold_ctf_daily}
\alias{hold_ctf_daily}
\title{Execute Batch API Download from BlackDiamond}
\usage{
hold_ctf_daily(bucket, api_keys, as_of = NULL)
}
\arguments{
\item{api_keys}{list with credentials}

\item{as_of}{optional parameter for as of date to pull holdings, leave
NULL for last trading day}
}
\value{
does not return data, holdings and transactions are stored in
  parquet files in S3
}
\description{
Run before ret_ctf_daily(), the holdings and transactions are needed to
  calculate the estimated return
}
