% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{hold_sec}
\alias{hold_sec}
\title{Download holdings from SEC EDGAR Database}
\usage{
hold_sec(
  bucket,
  dtc_name = NULL,
  user_email = "asotolongo@diversifiedtrust.com",
  save_to_db = TRUE,
  return_data = FALSE
)
}
\arguments{
\item{dtc_name}{leave `NULL` to download all, or enter a vector of
dtc_names to download specific funds}

\item{user_email}{need to provide an email address to download}

\item{save_to_db}{boolean to write to DTC's database, default is TRUE}

\item{return_data}{boolean to return data.frame of holdings, default is FALSE}
}
\description{
Download holdings from SEC EDGAR Database
}
