% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{imb_trail_perf_tbl}
\alias{imb_trail_perf_tbl}
\title{IMB Trailing Performance Table}
\usage{
imb_trail_perf_tbl(res, as_of)
}
\arguments{
\item{res}{from clean_asset_bench_rf()}

\item{as_of}{as of date for performance}
}
\value{
flextable with trailing performance
}
\description{
IMB Trailing Performance Table
}
