% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\name{latest_holdings}
\alias{latest_holdings}
\title{Subset the holdings file for the most recent date}
\usage{
latest_holdings(tbl_hold)
}
\arguments{
\item{tbl_hold}{data.frame of holdings}
}
\value{
tbl_hold with most recent holdings
}
\description{
Subset the holdings file for the most recent date
}
