% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\name{left_merge_flat}
\alias{left_merge_flat}
\title{Utility function to return left merge of x and y as data.frame}
\usage{
left_merge_flat(x, y, match_by, keep_x_dup_col = FALSE)
}
\arguments{
\item{x}{data.frame}

\item{y}{data.frame to form left join}

\item{match_by}{string to represent which field (column) to match}

\item{keep_x_dup_col}{boolean to keep duplicate columns in the x data.frame,
default is FALSE meaning common fields from y will replace counterparts in
x}
}
\value{
data.frame of x, y join
}
\description{
Utility function to return left merge of x and y as data.frame
}
