% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wrangle.R
\name{list_replace_null}
\alias{list_replace_null}
\title{Utility function to replace NULL values in list with NA}
\usage{
list_replace_null(x)
}
\arguments{
\item{x}{list}
}
\description{
Utility function to replace NULL values in list with NA
}
