% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\name{merge_qual}
\alias{merge_qual}
\title{Merge Qualitative Data with holdings}
\usage{
merge_qual(tbl_hold, list_data)
}
\arguments{
\item{tbl_hold}{data.frame with portfolio holdings}

\item{list_data}{list of data.frames with Qualitative Data}
}
\description{
Merge Qualitative Data with holdings
}
\note{
tables are merged based on DtcName field, left join
}
