% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{pvt_cf}
\alias{pvt_cf}
\title{Read Private Cash Flows from Excel}
\usage{
pvt_cf(wb, cf_sht = "cf", nav_sht = "nav", unsmooth = TRUE)
}
\arguments{
\item{wb}{workbook with cash flows and NAVs}

\item{cf_sht}{name of worksheet with cash flows}

\item{nav_sht}{name of worksheet with NAVs}

\item{unsmooth}{optional paramter to specify whether to unsmooth the
returns or not}
}
\value{
list with twr = time weighted returns xts and tvpi = vector of
  multiples
}
\description{
Read Private Cash Flows from Excel
}
