% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{pvt_cf_to_monthly}
\alias{pvt_cf_to_monthly}
\title{Put Monthly Public Indexes in Quarterly Private Returns}
\usage{
pvt_cf_to_monthly(bucket, res)
}
\arguments{
\item{bucket}{s3 filesystem}

\item{res}{from `twr_on_tvpi()` function}
}
\value{
xts with monthly returns
}
\description{
Put Monthly Public Indexes in Quarterly Private Returns
}
