% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\name{qual_roll_up}
\alias{qual_roll_up}
\title{Roll up qualitative data}
\usage{
qual_roll_up(res, bucket, qual = NULL)
}
\arguments{
\item{res}{output from drill_down() function}

\item{bucket}{s3 file system}

\item{qual}{optional list of qualitative data, leave NULL to load defaults}
}
\value{
list with qualitative data merged and total slot for totals
}
\description{
Roll up qualitative data
}
