% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{read_hold}
\alias{read_hold}
\title{Read holdings}
\usage{
read_hold(ids, bucket, tbl_msl = NULL)
}
\arguments{
\item{ids}{vector of IDs, can be Ticker, Cusip, DtcName, or any ID field in
the MSL}

\item{bucket}{S3FileSystem object}

\item{tbl_msl}{data.frame representing master security list}
}
\value{
list with holdings files
}
\description{
Read holdings
}
