% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket.R
\name{read_ret}
\alias{read_ret}
\title{Read Returns}
\usage{
read_ret(ids, bucket, tbl_msl = NULL)
}
\arguments{
\item{ids}{ticker, cusip, lei, dtc name, or identifier}

\item{bucket}{S3FileSystem object}

\item{tbl_msl}{option to pass through a custom master security list,
leave NULL for default MSL}
}
\value{
xts of time-series of ids, will warn if some ids are not found
}
\description{
Read Returns
}
