% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\name{remove_dup_dates}
\alias{remove_dup_dates}
\title{Remove Duplicate Dates in Holdings File}
\usage{
remove_dup_dates(tbl_hold, id = "Name")
}
\arguments{
\item{tbl_hold}{data.frame of holdings}

\item{id}{ID column name to use for duplicates, e.g., Name, Ticker, Cusip}
}
\value{
tbl_hold with any duplicate dates removed
}
\description{
Problems can arise if duplicate multiple holdings are written
  for the same date, e.g., if was the holdings were accidently gathered
  and rowbinded to the holdings file twice for the same date. This function
  is intended to search for those occurances by finding duplications of a
  date and ID and removing them.
}
