% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{ret_comp}
\alias{ret_comp}
\title{Update returns that require computational changes}
\usage{
ret_comp()
}
\value{
does not return data, updates various return records in the
  database
}
\description{
Update returns that require computational changes
}
\details{
for example cash plus 200 bps reads the returns of cash, adds 200 bps,
and then saves as a new data-field (return column)
}
