% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{ret_ctf_daily}
\alias{ret_ctf_daily}
\title{Update CTF Daily Returns}
\usage{
ret_ctf_daily(bucket, date_start = NULL, date_end = NULL)
}
\arguments{
\item{date_start}{first date for return update, defaultt is two trading
days prior to date_end}

\item{date_end}{lastest date for return update, default is last trading
day}
}
\value{
no data returned, row binds new returns to database "ctf-daily"
  record
}
\description{
Update CTF returns each day from Black Diamond bulk data
}
