% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{ret_ctf_daily_adj}
\alias{ret_ctf_daily_adj}
\title{Adjust daily CTF return estimates to land on month end returns}
\usage{
ret_ctf_daily_adj()
}
\value{
does not return data, updates "ctf-daily" record in database
}
\description{
Adjust daily CTF return estimates to land on month end returns
}
