% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{ret_ctf_monthly}
\alias{ret_ctf_monthly}
\title{Upload Month End CTF Reconciled Returns into the Database}
\usage{
ret_ctf_monthly(bucket, xl_path = NULL, skip = 4)
}
\arguments{
\item{xl_path}{filepath of excel workbook with returns, leave `NULL` for
default}

\item{skip}{leading header rows to skip, default is `4`}

\item{ac}{arcticDB datastore}
}
\value{
no data is returned, the entire history is read and stored in
  arcticDB
}
\description{
Upload Month End CTF Reconciled Returns into the Database
}
