% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{ret_date_info}
\alias{ret_date_info}
\title{Get Start and End Dates for each Asset in xts}
\usage{
ret_date_info(x)
}
\arguments{
\item{x}{xts}
}
\value{
data.frame with start and end dates for each column in xts
}
\description{
Get Start and End Dates for each Asset in xts
}
\examples{
data(assets)
ret_date_info(x)
}
