% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{ret_etf}
\alias{ret_etf}
\title{Update ETF returns from Factset}
\usage{
ret_etf(bucket, api_keys, ids = NULL, date_start = NULL, date_end = NULL)
}
\arguments{
\item{bucket}{S3 FileSystem created by create_bucket()}

\item{api_keys}{list of credentials}

\item{ids}{leave `NULL` to udpate all ETFs in the Master Security List,
or enter a vector to only update specific ETFs}

\item{date_start}{beginning date for update, if left `NULL` will
default to the last date of the existing data}

\item{date_end}{ending date for update, by default is yesterday}
}
\description{
Update ETF returns from Factset
}
