% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{ret_fred}
\alias{ret_fred}
\title{Update econ time-series from St. Louis FED (FRED)}
\usage{
ret_fred(bucket)
}
\value{
does not return data, updates "fred-monthly" record in database
}
\description{
Update econ time-series from St. Louis FED (FRED)
}
