% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{ret_hfr_index}
\alias{ret_hfr_index}
\title{Update HFRI Return index from csv file}
\usage{
ret_hfr_index(file_nm)
}
\arguments{
\item{file_nm}{full file name of csv file}
}
\value{
does not return data, updates "hfr-index" record in the database
}
\description{
Update HFRI Return index from csv file
}
\seealso{
\code{\link{read_hfr_index}}
}
