% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{ret_model}
\alias{ret_model}
\title{Update returns of models}
\usage{
ret_model(bucket, api_keys, dtc_name = NULL, months_back = 1)
}
\arguments{
\item{dtc_name}{option to specify specific models to update, leave NULL
to update all models}

\item{months_back}{integer representing how many months back to update
returns for, default is 1}
}
\description{
Update returns of models
}
