% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{ret_mutual_fund}
\alias{ret_mutual_fund}
\title{Update index returns from Factset}
\usage{
ret_mutual_fund(bucket, api_keys, ids = NULL, days_back = 1)
}
\arguments{
\item{bucket}{S3 FileSystem created by create_bucket()}

\item{api_keys}{list of credentials}

\item{ids}{leave NULL to update all mutual funds, or enter specific
mutual fund, note if ids are entered they must be in the MSL}

\item{days_back}{how many days to download}
}
\description{
Update index returns from Factset
}
