% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{ret_stock}
\alias{ret_stock}
\title{Update stock returns}
\usage{
ret_stock(
  bucket,
  api_keys,
  ids = NULL,
  date_start = NULL,
  date_end = NULL,
  freq = "D",
  geo = c("us", "intl")
)
}
\arguments{
\item{ids}{optional parameter to only update certain stocks, leave NULL
to update all stocks in the MSL}

\item{date_start}{starting date for new returns, default is NULL which
will find most recent date of existing returns to start}

\item{date_end}{most recent date for new returns, default is Sys.Date()}

\item{freq}{"D" for daily}

\item{geo}{"us" for US Stocks or "intl" for international stocks}
}
\value{
does not return data, either updates "us-stock" or "intl-stock"
  records in database
}
\description{
Update stock returns
}
\seealso{
\code{\link{download_fs_global_prices}}
}
