% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{ret_workup}
\alias{ret_workup}
\title{read in manually uploaded returns from Excel}
\usage{
ret_workup(xl_file = "N:/Investment Team/DATABASES/CustomRet/workup.xlsx")
}
\arguments{
\item{xl_file}{full file path of excel workbook}
}
\value{
does not return data, updates "workup" record in database
}
\description{
read in manually uploaded returns from Excel
}
