% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{roll_beta}
\alias{roll_beta}
\title{Rolling Beta to Common Benchmark}
\usage{
roll_beta(x, b, rf, n)
}
\arguments{
\item{x}{xts of asset returns}

\item{b}{xts of benchmark returns (one benchmark)}

\item{rf}{xts of risk-free}

\item{n}{number of periods for rolling window}
}
\description{
Rolling Beta to Common Benchmark
}
\details{
Trailing beta to benchmark is calculated for `n` periods. If the data
are monthly then `n` set to 36 is 36 months. The rolling period includes
the current period: e.g., for monthly returns with `n` = 3 on June will
include April, May, and June returns.
}
