% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{roll_down_vol}
\alias{roll_down_vol}
\title{Rolling Downside Volatility}
\usage{
roll_down_vol(x, n, b = NULL, freq = "days")
}
\arguments{
\item{x}{xts of returns}

\item{n}{number of periods for rolling calc window, see details}

\item{b}{optional benchmark to transform to active return}

\item{freq}{string: days, weeks, months, quarters, years}
}
\description{
Rolling Downside Volatility
}
\details{
If frequency is "months" and n = 12, a rolling 12 month volatility when
the asset is down is calculated. If `b` is specified then the calculation
will be the volatility of the active return to the benchmark. Note frequency
will not change the frequency, it is used to annaulize the volatility.
The rolling period includes the current period: e.g.,
for monthly returns with `n` = 3 on June will include April, May, and
June returns.
}
