% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{roll_r2}
\alias{roll_r2}
\title{Rolling R-squared to Common Benchmark}
\usage{
roll_r2(x, b, n)
}
\arguments{
\item{x}{xts of asset returns}

\item{b}{xts of benchmark return (one benchmark)}

\item{n}{number of periods for rolling window.}
}
\description{
Rolling R-squared to Common Benchmark
}
\details{
Trailing r^2 to benchmark is calculated for `n` periods. If the data
are monthly then `n` set to 36 is 36 months. The rolling period includes
the current period: e.g., for monthly returns with `n` = 3 on June will
include April, May, and June returns.
}
