% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{roll_ret}
\alias{roll_ret}
\title{Rolling trailing returns}
\usage{
roll_ret(x, n, b = NULL, freq = "days")
}
\arguments{
\item{x}{xts of returns}

\item{n}{number of periods for rolling calc window, see details}

\item{b}{optional benchmark to transform to active return}

\item{freq}{string: days, weeks, months, quarters, years}
}
\description{
Rolling trailing returns
}
\details{
If frequency is "months" and n = 12, a rolling 12 month return is
calculated. Note frequency will not change the frequency, it should
match the return periodicity. Frequency is used to determine if the
returns should be annualized, i.e., if the period is longer than 1 year.
Rolling return periods for less than 1 year will not be annualized.
The rolling period includes the current period: e.g.,
for monthly returns with `n` = 3 on June will include April, May, and
June returns.
}
