% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{roll_style}
\alias{roll_style}
\title{Rolling Style Analysis}
\usage{
roll_style(x, b, n)
}
\arguments{
\item{x}{xts of asset returns to explain}

\item{b}{xts of index or factor returns used to track `x`}

\item{n}{number of observations in rolling period}
}
\description{
Rolling Style Analysis
}
