% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{te_min_qp}
\alias{te_min_qp}
\title{Quadratic solver wrapper for minimizing tracking error}
\usage{
te_min_qp(fund, fact)
}
\arguments{
\item{fund}{xts of asset to replicate / minimize TE to}

\item{fact}{xts of factors / assets to use to try and replicate `fund`}
}
\value{
results from `quadprog::solve.QP`
}
\description{
Quadratic solver wrapper for minimizing tracking error
}
