% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{twr_on_tvpi}
\alias{twr_on_tvpi}
\title{Match Product of TWR to TVPI}
\usage{
twr_on_tvpi(res)
}
\arguments{
\item{res}{output of `pvt_cf()` function}
}
\value{
res input list with twr_adj added to represent adjusted time
  weighted returns that equal the TVPI multiple
}
\description{
Match Product of TWR to TVPI
}
