% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{update_ps_macro_select}
\alias{update_ps_macro_select}
\title{Download Macro Select Workbook and Save to Library}
\usage{
update_ps_macro_select(bucket, wb, is_us = TRUE)
}
\arguments{
\item{bucket}{s3 file system}

\item{wb}{file location of workbook}

\item{is_us}{TRUE for Russell 3000, FALSE for MSCI ACWI}
}
\description{
Download Macro Select Workbook and Save to Library
}
