% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time_series.R
\name{xts_eo_month}
\alias{xts_eo_month}
\title{Convert xts index to end of month dates}
\usage{
xts_eo_month(x)
}
\arguments{
\item{x}{xts object}
}
\value{
xts object with end of month dates as index
}
\description{
Convert xts index to end of month dates
}
