% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\name{drill_down}
\alias{drill_down}
\title{Drilldown into holdings table}
\usage{
drill_down(bucket, tbl_hold)
}
\arguments{
\item{bucket}{S3 file system}

\item{tbl_hold}{data.frame of portfolio holdings}
}
\value{
list with different layers
}
\description{
Drilldown into holdings table
}
